cleanup of unnecessary styles and translation labels + better folder structure
