+++ frontend/src/views/AvailabilityView/components/AvailabilitySettings/components/AvailabilityCopyDropdown.vue
