+++ frontend/src/views/AvailabilityView/components/AvailabilitySettings/components/AvailabilitySelect.vue
